requirements.minimal.txt
Django==1.7.9
dj-static==0.0.6
gunicorn==18.0
static3==0.5.1
wsgiref==0.1.2
dj-database-url==0.2.2
django-browserid==0.11.1
requests==2.5.3
djangorestframework==3.1.2
geopy==1.9.1
django-cors-headers==1.0.0
django-csp==2.0.3
Markdown==2.6.1
slumber==0.7.1

# For tests to run...
mock==1.0.1
httmock==1.2.2
factory_boy==2.7.0
